da hoan thanh auth service 28/05

+++ Backend/E-commerce/E-commerce/Services/notification/src/main/java/com/local/ecommerce/kafka/order/Product.java
@@ -7,6 +7,6 @@ public record Product(
         String name,
         String description,
         BigDecimal price,
-        double amount
+        BigDecimal amount
 ) {
 }
